Factor out LocalModRegistry
